feat(city): add exceptions handle in the method of removing cities from the city repository

+++ src/main/java/com/studies/algafood/infrastructure/repository/CityRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.studies.algafood.domain.model.City;
 import com.studies.algafood.domain.repository.CityRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,8 +34,12 @@ public class CityRepositoryImpl implements CityRepository {
 
     @Override
     @Transactional
-    public void remove(City city) {
-        city = find(city.getId());
+    public void remove(Long id) {
+        City city = find(id);
+
+        if (city == null){
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(city);
     }
 }
